requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asn1crypto=1.4.0=py_0
ca-certificates=2021.4.13=haa95532_1
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.13.2=py27hcfb25f9_0
chardet=3.0.4=py27_1003
click=7.1.2=pyhd3eb1b0_0
cryptography=2.8=py27hcfb25f9_0
enum34=1.1.6=py27_1
idna=2.10=pyhd3eb1b0_0
ipaddress=1.0.23=py_0
openssl=1.1.1e=h0c8e037_0
pip=19.3.1=py27_0
pycparser=2.20=py_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pysocks=1.7.1=py27_0
python=2.7.18=hfb89ab9_0
python-dotenv=0.17.0=pyhd8ed1ab_0
pytz=2021.1=pyhd3eb1b0_0
requests=2.25.1=pyhd3eb1b0_0
setuptools=44.0.0=py27_0
six=1.15.0=pyhd3eb1b0_0
sqlite=3.30.1=h0c8e037_0
urllib3=1.25.7=py27_0
vc=9=h7299396_1
vs2008_runtime=9.00.30729.1=hfaea7d5_1
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py27_0
wincertstore=0.2=py27hf04cefb_0
